refactor: harden result caching and remove dead code after audit
Pre-submission audit of the reproduction scripts. Correctness fixes:

- RoarGraph search results were cached in one filename shared by every
  Recall@K target, so a multi-K run (Figure 4 sweeps k=1/10/100) silently
  served the first K's numbers for the rest. The result/CSV files now carry
  target_k. This also repairs the merge step, which looked for a per-K CSV
  that was never written, leaving RoarGraph out of the comparison tables.
- Base indices and mined shortcut sets were cached under names that ignored
  the parameters that determine their contents, so re-running with a
  different --m / --ef-construction / --budget / --train-ef / --top-k
  silently reused stale artifacts. Cache names now include them
  (common.base_index_filename / common.shortcut_filename).

Cleanups:

- Consolidate three near-duplicate recall interpolation routines into
  common.qps_at_recall and common.qps_at_recall_pareto; verified to
  reproduce the previous Table 5 numbers exactly on real COCO data.
- Drop dead code: ours_utils.compare_results, ours_utils.resolve_dataset_path,
  Ours_Miner._filter_and_save_cpp, the unused evaluate_all_k return value,
  and the unused sorted_labels return from reorder_dataset_by_clustering.
- Rename Ours_Miner.train_frequency_shortcuts to mine_shortcuts (the
  algorithm is difference-based, not frequency-based, per Sec. 3.4) and drop
  its unused cluster_labels parameter.
- Fast-path the ground-truth padding loop for the rectangular arrays the
  benchmarks actually provide.
- Cast to size_t in the candidate-budget comparisons; ours_backend.cpp now
  builds clean under -Wall -Wextra.
- Constrain --num-runs so 0 fails in argparse instead of a later NameError.

+++ src/ours_backend.cpp
@@ -119,7 +119,7 @@ void filter_and_save_cpp(
         std::vector<int> candidates;
         for (int j = start; j < end; j++) candidates.push_back(r_indices(j));
 
-        if (candidates.size() > budget * 2) {
+        if (candidates.size() > static_cast<size_t>(budget) * 2) {
              std::sort(candidates.begin(), candidates.end(), [&](int a, int b) {
                  float dist_a = compute_ip(&r_data(i, 0), &r_data(a, 0), dim);
                  float dist_b = compute_ip(&r_data(i, 0), &r_data(b, 0), dim);
@@ -136,7 +136,7 @@ void filter_and_save_cpp(
 
         std::vector<int> selected;
         for (auto& p : cand_dists) {
-            if (selected.size() >= budget) break;
+            if (selected.size() >= static_cast<size_t>(budget)) break;
             int cand_idx = p.second;
             float dist_to_base = p.first;
 
@@ -285,7 +285,7 @@ double mine_and_filter_cpp(
                 return a.first > b.first; // descending inner product
             });
 
-            if (cand_dists.size() > budget * 2) {
+            if (cand_dists.size() > static_cast<size_t>(budget) * 2) {
                 cand_dists.resize(budget * 2);
             }
 
@@ -296,7 +296,7 @@ double mine_and_filter_cpp(
                 // RNG heuristic: keep a candidate only if no previously
                 // accepted neighbor lies strictly closer to it than to i.
                 for (auto& p : cand_dists) {
-                    if (selected.size() >= budget) break;
+                    if (selected.size() >= static_cast<size_t>(budget)) break;
                     int cand_idx = p.second;
                     float dist_to_base = p.first;
 
@@ -313,7 +313,7 @@ double mine_and_filter_cpp(
             } else {
                 // Ablation: ignore RNG diversity, keep the closest `budget`.
                 for (auto& p : cand_dists) {
-                    if (selected.size() >= budget) break;
+                    if (selected.size() >= static_cast<size_t>(budget)) break;
                     selected.push_back(p.second);
                 }
             }
